Move energy_calculations out of /utils
