+++ src/application/energy_calculation.ts
